requirements-direct.txt
# These are the direct dependencies of CourtListener only, removing any
# subdependencies.
#
# Note that we use ~= in many places to note that minor versions are allowed
# here, while in requirements.txt we always use == to lock things down properly

ansible
beautifulsoup4~=4.8.0
# Required for django-storage
boto3~=1.9.0
# Before upgrading, note: https://github.com/celery/celery/issues/4878
celery~=4.3.0
# Always do the most updated version of certifi, no version lock.
certifi
# Needed for get_schema_view view (https://www.django-rest-framework.org/coreapi/schemas/#adding-a-view-with-get_schema_view)
coreapi~=2.3.0
# Unfortunately, higher versions emit an insane Python 2.7 error EVERYWHERE.
# Once we have py3, we can upgrade: https://stackoverflow.com/q/51778764/64911
cryptography==2.2.2
defusedxml~=0.6.0
disposable-email-domains
Django~=1.11.0
# We have the last version of this that supports py2.
django-cors-headers~=3.0.0

# XXX Items below here need vetting for Py3.
django-filter==1.1
django-localflavor==1.1
django-mathfilters
django-markdown-deux
django-ratelimit~=2.0.0
django-redis-cache~=2.0.0
djangorestframework==3.6.2
djangorestframework-filters~=0.11.0
django-rest-swagger==2.1.2
django-storages~=1.7.0
djangorestframework-xml==1.4.0
docutils==0.12
drf-dynamic-fields
enum34==1.1.2
eyeD3==0.7.11
feedparser==5.2.1
geonamescache==0.20
gevent<1.3
google-cloud-storage==1.6.0
google-api-python-client==1.6.4
hiredis==0.2.0
internetarchive
ipaddress==1.0.16
ipython==3.2.0
itypes==1.1.0
# Kombu 4.6.5 is bad, and its authors should feel bad. Pin Kombu at a lower
# version, which actually should work. See:
# https://github.com/celery/kombu/issues/1063
kombu==4.6.4
logilab-common==0.63.2
lxml~=4.0
Markdown~=2.6
markdown2~=2.3.0
mock
ndg-httpsclient==0.4.0
networkx==1.10
nose
openapi-codec==1.3.1
pandas==0.18.1
Pillow~=3.3
psycopg2>=2.7,<2.8 --no-binary psycopg2
pycparser==2.14
pydot==1.1.0
pyinotify==0.9.6; sys_platform != 'darwin'
pyOpenSSL
pyparsing==2.4.2
PyPDF2==1.26.0
python-dateutil==2.5.0
python-igraph
python-mimeparse==0.1.4
pytz
redis~=3.2.0
requests==2.20.0
scorched==0.11.0
seal_rookery
selenium~=3.0
sentry-sdk~=0.0
simplejson
six
sqlparse==0.1.15
stripe==1.22.3
timeout-decorator
tldextract
usaddress
wsgiref==0.1.2
xlrd
unidecode
